Allow non-exclusive UDP comm i.e. fix multicasting
- UdpWrapperTests.cs is modified to test multicasting

- UdpWrapper.cs has undergone the most changes. The passed UdpClient
is expected to not be connected. It will send its messages to
`remote`. If `exclusiveConnection`, messages from anyone other than
`remote` will be ignored, while if it is false, all messages will be
received. The latter behavior is necessary for multicasting, which
is what this commit was supposed to fix.

- In InvitationManager.cs and PlayerListCoordinator.cs all calls to
the UdpWrapper constructor are modified to the new interface